+++ lib/src/client/quantum_fetch_imp.dart
@@ -13,102 +13,136 @@ class QuantumFetch implements IQuantumFetch {
   }
 
   @override
-  Future<HttpResponse<T, T>> get<T>(String path,
-      {Map<String, String> headers = const {},
-      T Function(Map<String, dynamic>)? decoder,
-      OnProgress? onProgress,
-      String? dataNode = "data"}) async {
-    final response = await getRaw(path, onProgress: onProgress);
+  Future<APIResponse<T>> get<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    T Function(Map<String, dynamic>)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
+    final response =
+        await getRaw(path, onProgress: onProgress, headers: headers);
     return HttpResponse<T, T>.fromDioResponse(
-        response, (json) => decoder?.call(json) ?? json as T, dataNode);
+        response, (json) => decoder?.call(json) ?? json as T, dataNode, config);
   }
 
   Future<Response<dynamic>> getRaw(String path,
-      {OnProgress? onProgress}) async {
+      {OnProgress? onProgress, Map<String, dynamic> headers = const {}}) async {
     final dio = await instance;
     final response = await dio.get(path,
+        options: Options(headers: headers),
         onReceiveProgress: ((count, total) =>
             onProgress?.call(total ~/ count)));
+    return response;
+  }
 
+  Future<Response<dynamic>> postRaw(String path,
+      {OnProgress? onProgress,
+      Map<String, dynamic> data = const {},
+      Map<String, dynamic> headers = const {}}) async {
+    final dio = await instance;
+    final response = await dio.post(path,
+        data: data,
+        options: Options(headers: headers),
+        onSendProgress: ((count, total) => onProgress?.call(total ~/ count)));
+    return response;
+  }
+
+  Future<Response<dynamic>> patchRaw(String path,
+      {OnProgress? onProgress,
+      Map<String, dynamic> data = const {},
+      Map<String, dynamic> headers = const {}}) async {
+    final dio = await instance;
+    final response = await dio.patch(path,
+        data: data,
+        options: Options(headers: headers),
+        onSendProgress: ((count, total) => onProgress?.call(total ~/ count)));
     return response;
   }
 
   @override
-  Future<HttpResponse<List<T>, T>> getList<T>(String path,
-      {Map<String, String> headers = const {},
-      T Function(Map<String, dynamic>)? decoder,
-      OnProgress? onProgress,
-      String? dataNode = "data"}) async {
-    final response = await getRaw(path, onProgress: onProgress);
+  Future<APIResponseList<T>> getList<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    T Function(Map<String, dynamic>)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
+    final response =
+        await getRaw(path, onProgress: onProgress, headers: headers);
     return HttpResponse<List<T>, T>.fromDioResponse(
-        response, (json) => decoder?.call(json) ?? json as T, dataNode);
+        response, (json) => decoder?.call(json) ?? json as T, dataNode, config);
   }
 
   @override
-  Future<HttpResponse<T, T>> post<T>(String path,
-          {Map<String, String> headers = const {},
-          Map<String, dynamic> body = const {},
-          T Function(Map<String, dynamic>)? decoder,
-          OnProgress? onProgress,
-          String? dataNode = "data"}) async =>
-      postAndGetList<T, T>(path,
-          headers: headers,
-          body: body,
-          decoder: decoder,
-          dataNode: dataNode,
-          onProgress: onProgress);
-  @override
-  Future<HttpResponse<T, K>> postAndGetList<T, K>(String path,
-      {Map<String, String> headers = const {},
-      Map<String, dynamic> body = const {},
-      K Function(Map<String, dynamic> p1)? decoder,
-      OnProgress? onProgress,
-      String? dataNode = "data"}) async {
-    final dio = await instance;
-    final response = await dio.post(path,
-        data: body,
-        onSendProgress: ((count, total) =>
-            onProgress?.call(calculateProgress(count, total))));
-    return HttpResponse<T, K>.fromDioResponse(
-        response, (json) => decoder?.call(json) ?? json as K, dataNode);
+  Future<APIResponse<T>> post<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    Map<String, dynamic> body = const {},
+    T Function(Map<String, dynamic>)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
+    final response = await postRaw(path,
+        data: body, headers: headers, onProgress: onProgress);
+    return HttpResponse<T, T>.fromDioResponse(
+        response, (json) => decoder?.call(json) ?? json as T, dataNode, config);
   }
 
   @override
-  Future<HttpResponse<T, T>> patch<T>(String path,
-          {Map<String, String> headers = const {},
-          Map<String, dynamic> body = const {},
-          T Function(Map<String, dynamic>)? decoder,
-          OnProgress? onProgress,
-          String? dataNode = "data"}) async =>
-      patchAndGetList<T, T>(path,
-          headers: headers,
-          body: body,
-          decoder: decoder,
-          dataNode: dataNode,
-          onProgress: onProgress);
-
-  @override
-  Future<HttpResponse<T, K>> patchAndGetList<T, K>(String path,
-      {Map<String, String> headers = const {},
-      Map<String, dynamic> body = const {},
-      K Function(Map<String, dynamic> p1)? decoder,
-      OnProgress? onProgress,
-      String? dataNode = "data"}) async {
-    final dio = await instance;
-    final response = await dio.patch(path,
-        data: body,
-        onSendProgress: ((count, total) => onProgress?.call(total ~/ count)));
-    return HttpResponse<T, K>.fromDioResponse(
-        response, (json) => decoder?.call(json) ?? json as K, dataNode);
+  Future<APIResponseList<T>> postAndGetList<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    Map<String, dynamic> body = const {},
+    T Function(Map<String, dynamic> p1)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
+    final response = await postRaw(path,
+        data: body, headers: headers, onProgress: onProgress);
+    return HttpResponse<List<T>, T>.fromDioResponse(
+        response, (json) => decoder?.call(json) ?? json as T, dataNode, config);
   }
 
   @override
-  Future<HttpResponse<T, T>> delete<T>(String path,
-      {Map<String, String> headers = const {},
-      Map<String, dynamic> body = const {},
-      T Function(Map<String, dynamic>)? decoder,
-      OnProgress? onProgress,
-      String? dataNode = "data"}) async {
+  Future<APIResponse<T>> patch<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    Map<String, dynamic> body = const {},
+    T Function(Map<String, dynamic>)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
+    final response = await patchRaw(path,
+        data: body, headers: headers, onProgress: onProgress);
+    return HttpResponse<T, T>.fromDioResponse(
+        response, (json) => decoder?.call(json) ?? json as T, dataNode, config);
+  }
+
+  @override
+  Future<APIResponseList<T>> patchAndGetList<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    Map<String, dynamic> body = const {},
+    T Function(Map<String, dynamic> p1)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
+    final response = await patchRaw(path,
+        data: body, headers: headers, onProgress: onProgress);
+    return HttpResponse<List<T>, T>.fromDioResponse(
+        response, (json) => decoder?.call(json) ?? json as T, dataNode, config);
+  }
+
+  @override
+  Future<APIResponse<T>> delete<T>(
+    String path, {
+    Map<String, String> headers = const {},
+    Map<String, dynamic> body = const {},
+    T Function(Map<String, dynamic>)? decoder,
+    OnProgress? onProgress,
+    JsonResponseNode? dataNode,
+  }) async {
     final dio = await instance;
     final response = await dio.delete(path, data: body);
     final responseBody = response.data ?? {};
@@ -128,16 +162,16 @@ class QuantumFetch implements IQuantumFetch {
     return Dio(
       BaseOptions(
           baseUrl: config.baseUrl,
-          connectTimeout: const Duration(seconds: 10),
-          receiveTimeout: const Duration(seconds: 10),
+          connectTimeout: Duration(seconds: config.connectTimeout),
+          receiveTimeout: Duration(seconds: config.receiveTimeout),
           validateStatus: (d) => true,
           headers: await getDefaultHeaders()),
     )..interceptors.addAll([
-        cacheIntercepter(),
         LogInterceptor(
             requestBody: true, responseBody: true, responseHeader: false),
         RequestBodyIntercepter(),
-        ...config.interceptors
+        ...config.interceptors,
+        cacheIntercepter(),
       ]);
   }
 }