+++ lib/src/response/data_node.dart
@@ -1,0 +1,4 @@
+class JsonResponseNode {
+  final String? nodeName;
+  JsonResponseNode(this.nodeName);
+}