+++ lib/src/response/response.dart
@@ -1,21 +1,33 @@
 import 'package:dio/dio.dart';
+import 'package:quantum_fetch/quantum_fetch.dart';
 
 class HttpResponse<T, K> {
   T? data;
   int? statusCode;
   String? message;
   bool success;
+
   HttpResponse({
     this.data,
     this.statusCode,
     this.message,
     required this.success,
   });
-  factory HttpResponse.fromDioResponse(Response response,
-      K Function(Map<String, dynamic>) decoder, String? rootNode) {
+  factory HttpResponse.fromDioResponse(
+      Response response,
+      K Function(Map<String, dynamic>) decoder,
+      JsonResponseNode? node,
+      final QuantumFetchConfig globalFetchConfig,
+      {List<int> validStatusCodes = const [200, 201, 204]}) {
+    ///check if there is modification from local request else check from global fetch config for root node modifications
+    final rootNode =
+        node != null ? node.nodeName : globalFetchConfig.dataNode.nodeName;
+    final successNode = globalFetchConfig.successNode.nodeName;
     final json = response.data;
     T? data;
-    final ok = json['ok'] as bool? ?? false;
+    final ok = successNode != null
+        ? (json[successNode] as bool? ?? false)
+        : validStatusCodes.contains(response.statusCode);
     final payloadData = rootNode == null ? json : json[rootNode];
     if (ok) {
       if (T == List<K>) {
@@ -30,7 +42,7 @@ class HttpResponse<T, K> {
         data: data,
         statusCode: response.statusCode,
         message: errorMessageDecoder(json),
-        success: json['ok'] as bool? ?? false);
+        success: ok);
   }
 }