+++ CHANGELOG.md
@@ -1,3 +1,16 @@
-## 0.0.1
+## 0.0.6
+* Fetch config global rootnode issue fixed
 
-* TODO: Describe initial release.
+## 0.0.5
+* Root node customization from fetch config and from request level function as well
+## 0.0.3
+* Added getList method
+## 0.0.2
+## 0.0.4
+* Added PostAndGetList ,PatchAndGetList features
+
+* Added cache interceptor
+## 0.0.1
+* first version
+
+